Flattening Code Struct
