+++ magia/__init__.py
@@ -10,9 +10,11 @@ from importlib.util import find_spec
 from . import factory
 from .bundle import Bundle, BundleSpec, BundleType
 from .constant import Constant
+from .elaborator import Elaborator
+from .io_ports import IOPorts
 from .io_signal import Input, Output
 from .memory import Memory
-from .module import Elaborator, Instance, IOPorts, Module, VerilogWrapper
+from .module import Instance, Module, VerilogWrapper
 from .register import Register
 from .signals import Signal
